Decouple reflection buffer convolution code from SSR-specific code
