+++ Include/Althea/ReflectionBuffer.h
@@ -14,7 +14,6 @@
 #include "Sampler.h"
 #include "ComputePipeline.h"
 #include "PerFrameResources.h"
-#include "ReflectionBuffer.h"
 
 #include <vulkan/vulkan.h>
 
@@ -23,33 +22,33 @@
 namespace AltheaEngine {
 class Application;
 
-class ALTHEA_API ScreenSpaceReflection {
+class ReflectionBuffer {
 public:
-  ScreenSpaceReflection() = default;
-  ScreenSpaceReflection(
+  ReflectionBuffer() = default;
+  ReflectionBuffer(
       const Application& app,
-      VkCommandBuffer commandBuffer,
-      VkDescriptorSetLayout globalSetLayout,
-      const GBufferResources& gBuffer);
+      VkCommandBuffer commandBuffer);
   void transitionToAttachment(VkCommandBuffer commandBuffer);
-  void captureReflection(
-      const Application& app,
-      VkCommandBuffer commandBuffer,
-      VkDescriptorSet globalSet,
-      const FrameContext& context);
   void convolveReflectionBuffer(
       const Application& app,
       VkCommandBuffer commandBuffer,
       const FrameContext& context);
-
   void bindTexture(ResourcesAssignment& assignment) const;
 
+  const ImageView& getReflectionBufferTargetView() const { return this->_mipViews[0]; }
 private:
-  ReflectionBuffer _reflectionBuffer;
-  
-  std::unique_ptr<RenderPass> _pReflectionPass;
-  std::unique_ptr<DescriptorSetAllocator> _pGBufferMaterialAllocator;
-  std::unique_ptr<Material> _pGBufferMaterial;
-  FrameBuffer _reflectionFrameBuffer;
+  std::unique_ptr<ComputePipeline> _pConvolutionPass;
+  std::unique_ptr<DescriptorSetAllocator> _pConvolutionMaterialAllocator;
+  std::vector<Material> _convolutionMaterials;
+
+  // Entire reflection buffer resource
+  // The view contains all the mips together
+  ImageResource _reflectionBuffer;
+
+  // Individual mip views of the reflection buffer mips
+  std::vector<ImageView> _mipViews;
+  // The sampler to use when viewing a single mip of the
+  // reflection buffer.
+  Sampler _mipSampler;
 };
 } // namespace AltheaEngine
